Extract Transaction, ioctl_guard! and PF helpers into individual modules
